DYADCOUNT() macro now uses and returns unsigned long int types internally. This prevents overflows, particularly on undirected networks with 46342<=n<=65536, whose dyad counts fit into a signed int, but the intermediate calculation can overflow.
git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/statnet_commons/ergm/branches/3.1@12242 9cb281f4-553f-406f-8d7f-11f072bbfe46

+++ src/edgetree.h
@@ -17,7 +17,7 @@
 
 #define MIN(a,b) ((a)<(b) ? (a) : (b))
 #define MAX(a,b) ((a)<(b) ? (b) : (a))
-#define DYADCOUNT(nnodes, bipartite, directed) (bipartite? (nnodes-bipartite)*bipartite : (directed? nnodes*(nnodes-1) : (nnodes*(nnodes-1))/2))
+#define DYADCOUNT(nnodes, bipartite, directed) (bipartite? (unsigned long)(nnodes-bipartite)*(unsigned long)bipartite : (directed? (unsigned long)nnodes*(unsigned long)(nnodes-1) : (((unsigned long)nnodes*(unsigned long)(nnodes-1))/2)))
 
 /*typedef unsigned int Vertex;
 typedef unsigned int Edge; */
